recipe-server/requirements/prod.txt
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
flask-marshmallow==0.10.1
marshmallow-sqlalchemy==0.21.0
gunicorn==20.0.4
gevent==1.4.0
flasgger==0.9.4